Restore Mark's commented out linkMode observer
Ref issue #3.

Adding geometry is expensive, so it would be better to be able to just
style the lines to be invisible.

A hacky way to do that is to just set stroke and fill to false; leaflet
doesn't seem to have a hide() method for features.

I suggest that you store a boolean vector `visible` as a reactive value
in the server scope. We'll then have the layer styling code depend on it
and I'll extend the setStyle method to do something sensible with it.