Start to remove calls to state()
This starts to adopt a convention where state is unpacked in
function arguments so that new parts of state can be accessed
easily by adding them to the function signature.

This way we can easily see which parts of the state are used
in a function without having to rewrite the calls to that
function every time we want to access more state.