Keep links data as a vector
We shouldn't return a matrix where other methods will expect a vector.
`fill` is also more understandable than doing weird stuff with the
Missing constructor.

We expect the client to call this sometimes with parameters that will
give a KeyError, but, particularly when loading a new dataset, seeing
these errors can be useful.